Correcciones docker

+++ backend/src/main/java/com/diegorojas/gestion_escolar_backend/domain/nota/DatosNota.java
@@ -2,11 +2,11 @@ package com.diegorojas.gestion_escolar_backend.domain.nota;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record DatosNota(
-        double valor,
+        BigDecimal valor,
         @NotNull
         Long alumno_id,
         @NotNull
